add ktor method for fetch data from fake server and setup config for change base url in product stash

+++ app/src/main/java/com/iamhessam/jsonplaceholder/data/remote/http/ktor/dto/CommentDTO.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CommentDTO(
-    var postId: Int,
-    var id: Int,
+    var postId: Long,
+    var id: Long,
     var name: String? = null,
     var email: String? = null,
     var body: String? = null
